pyproject.toml
[project]
name = "tgnews"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Dima Kuchin", email = "dima@mirable.io" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "redis>=6.4.0",
    "requests>=2.32.5",
]
